commit dos exercícios e desafios da aula 01

+++ exercicio_02.py
@@ -1,6 +1,4 @@
-# Crie um programa para retornar a soma de 2 valores digitados pelo usuário
-
-valor_1 = int(input("Digite um valor: "))
-valor_2 = int(input("Digite outro valor: "))
+valor_1 = float(input("Digite um valor: "))
+valor_2 = float(input("Digite outro valor: "))
 
 print(valor_1 + valor_2)
